+++ src/main/java/kangnamuniv/elibrary/service/LoanService.java
@@ -5,7 +5,7 @@ import kangnamuniv.elibrary.entity.Loan;
 import java.util.ArrayList;
 
 public interface LoanService {
-    Loan loanBook(Long userId, Long bookId);
+    Loan loanBook(Long userId, int bookId);
     Loan returnBook(Long loanId);
     ArrayList<Loan> findLoansByUser(Long userId);
 }
